Update specs --> specs2
Really?? I mean really?? This evening I have mostly been wanking with the build system and the dependencies. Instead of doing anything relevant to the actual tool.

I think it all started when intellij idea innocently wanted to update the scala plugin. I (what an ass) said yes. Therefore this evening I discovered that the specs runner is gone, so none of the tests are runnable in idea. Idea says that it remembers some run configurations for running tests which are now not applicable and therefore can't be run or changed or looked into for details on what the hell bit of code it was that previously did know how to run them. For I while I messed about with the plethora of traits and runner shit that are supposed to make it work (http://code.google.com/p/specs/wiki/RunningSpecs), without any luck. I decided against adding some extra random test dependencies that maybe might have given me a better chance of running the tests in idea, since -- with the entire cross-build scala versions, then cross-build sbt-versions, then test with different sbt-versions running different scala versions compiling code written for different scala versions debacle already giving me enough of a headache -- I really really can't see the point of adding more complexity into the mix to basically to able to run java command lines from idea to tell me if the exit value is 0 or not. Then I tried to google about rollbacking the idea scala plugin update, without any luck. Then I tried to see what was changed in the scala plugin, without it being at all informative to my situation. So could be the specs-runner just disappeared on it's own, who the hell knows. Anyway, then I looked into updating the specs-dependency to the latest scala (and self) version, only to find out the all the repositories have been moved to different urls, so none of the examples / repository-browsers could actually tell me where the hell I should look for the latest version. So I messed around with repository urls for a while, trying to figure out what was a snapshot repo and what was a release repo and other enterprise bullshit like that. Anyway, having found the latest specs version for the correct version of scala, it helped completely not at all. So, I tried a separate idea+sbt project with a specs2 test just to make sure that would work if I decided to port all the tests. Ok, finally something worked, it ran tests in both sbt and idea. That was so unexpected that I was slightly happy for a while, before realizing that 'yeah, wouldn't that be what was expected'. Anyway, then I decided to port all the specs to specs2. Since the addToSusVerb has apparently just completely disappeared, all the '"argh" should "jebus" in' - mumbo jumbo is now even more cryptic than before. Doesn't really matter, since it's never really added any value to the tests anyway.

I feel better for having written this.