Resolves canonical path in BaseReportWriter.relativeSource
We need the canonical path for the given src because our formattedSourcePaths are canonical.

The sources we iniitally store are non-canonical, provided directly through SBT's settings keys, which could
contain things like simlinks and '.' or '..' characters. This causes problems when trying to resolves sources
because we try to match against canonical formattedSourcePaths.

Misc.

- Bump SBT version
- Bump PGP plugin version