Change internal format from XML to plain text
Change format of `scoverage.coverage.xml` file (containing information about instrumented statements) from XML to plain test (and file name to `scoverage.coverage`).

In Scala version 2.13.0-M4 the dependency `scala-compiler` -> `scala-xml` was removed and since Scalac plugin cannot add its own dependencies, one of the solutions is to resign from XML format.

The code generating reports will still use `scala-xml` dependency. It's not a problem because it's executes in the context of the build tool (SBT, Maven or Gradle), not Scalac.

The side effect of this change is increased performance because writing and reading plain test files is much faster than XML files, especially when these files are large.