Don't do a full binary cross of the runtime.
Looking back at
https://github.com/scoverage/scalac-scoverage-plugin/commit/ac3db322c54dee18c967764d95f9904ff6d7e89f,
I'm not actually 100% sure this was accurate. While we do need to do a
fully binary publish of the actual compiler plugin, doing a full cross
of the runtime artifact causes some issues. Mainly I just don't believe
it's needed, but also then when publishing for other platforms, there is
odd behavior where you no longer get the sjs prefix that you'd want in
your artifacts published for JS. This change updates a few of the sbt
dependencies, moves to the unified slash syntax, and only applies the
full cross settings to the actual compiler plugin, not the runtime
artifact.