Update CI and release strategy (#315)
This simplifies the CI a bit and also moves to a process that I'm a bit
more familiar with. Mainly, everything for releases should now be able
to be handled fully by just creating a tag. I've found this to be much
easier to manage.

I also removed the `state.yml` since it's currently not working, and I'm
not sure how I feel about having something that will auto-close issues.