Add in release steps. (#331)
Just to document how the release is done I've gone ahead and added a
release section into a CONTRIBUTING.md file. I've also migrated the
release notes in the README to the corresponding releases on GitHub to
clean up the README a bit. Same with the license, since this is
displayed on GitHub and also included in the repo, I just removed it
from the README. NOTE the license still fully remains the same.