Add in scalafmt and scalafix.
I was going to wait to do this, but I was starting on something else in
here and I kept hitting format by habit since I'm honestly just used to
always having it. I think it's safe to say that most Scala developers
will be used to the default style that scalafmt gives you, so I'd like
to propose that we add this in along with scalafix to organize imports.

One huge concern is always that changes like this make it hard to get a
good `git blame`, however with modern git, this really isn't an issue
anymore. You can read more about his here:
https://www.moxio.com/blog/43/ignoring-bulk-change-commits-with-git-blame

I recently used the strategy outlined in that article in the
sbt-scoverage plugin:
https://github.com/scoverage/sbt-scoverage/pull/351 and it seems to work
quite well.