Migrate from scalatest to munit.
This is an opinionated change, but I'm a firm believer in munit.
There were some test issues previously that were pretty hard to debug
due to the nasty diffs that scalatest gives you compared to munit. Plus
when doing the migration I actually found some mistakes in the tests that
were false positives which switching from === to `assertEquals` unveiled.

It's a pretty straightforward migration.