Incremental coverage support
This makes scoverage work with incremental compilation.
Instead of starting from scratch we reuse the data in
`scoverage.coverage` for files that are not being compiled in this
compilation unit (because zinc skipped them).

We read the existing `scoverage.coverage` and merge it with the new
one being generated.

We make sure that statement ids by starting from the max id in the
existing coverage.

This PR also refactors the test code a little bit to be able
to test for incremental compilation.